requirements.txt
# ============================================
# Automated Enterprise Loan Approval System
# Python Dependencies
# ============================================
# Install all: pip install -r requirements.txt
# ============================================

# --- Data Processing ---
pandas==2.2.3
numpy==2.2.3

# --- Machine Learning & Analytics ---
scikit-learn==1.6.1
xgboost==2.1.4
joblib==1.4.2
plotly==6.0.0

# --- Testing ---
pytest==8.3.5
httpx==0.28.1

# --- Visualization ---
matplotlib==3.10.1
seaborn==0.13.2

# --- API & Web ---
fastapi==0.115.12
uvicorn==0.34.0
streamlit==1.42.2
requests==2.32.3

# --- Document Generation ---
fpdf==1.7.2

# --- Security ---
bcrypt==4.2.0

# --- Environment Variables ---
python-dotenv==1.1.0

# --- LLM (OpenAI) ---
openai==1.68.0

# --- Jupyter Notebooks (optional) ---
jupyter==1.1.1
